Removed api_helper.rb
