+++ app/serializers/avatar_obj_serializer.rb
@@ -1,5 +1,7 @@
 class AvatarObjSerializer < BaseSerializer
   attribute :avatar do
-    avatar_href_link(object)
+    uri = nil
+    uri = "/#{object.class.name.pluralize.underscore}/#{object.id}/avatars/#{object.avatar.id}" unless object.avatar.nil?
+    uri
   end
 end
